+++ doc/version_history.rst
@@ -15,6 +15,7 @@ v0.3.0
 * Moved BaseDevice and MockDevice from ts.ess.controller to ts.ess.common.
 * Added a unit test for the config schema.
 * Moved most of the command handler code and the socket server unit test from ts.ess.controller to ts.ess.common.
+* Added tests for all supported devices in the test class for the mock control handler.
 
 Requires: